langchain-langgraph-edenmarco/section-5-function-calling/pyproject.toml
[project]
name = "section-5-function-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
]
